Move helper functions to non-view file
